requirements.txt
requests==2.28.2
pyinstaller==5.7.0
tqdm==4.65.0
PyYAML==6.0
PyExecJS==1.5.1
bs4==0.0.1
prettytable==0.7.2
youtube-dl==2021.12.17
lxml==4.9.3
m3u8==4.1.0
pycryptodome==3.20.0
FFmpeg==1.4